# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "E-Commerce Platform API",
    "version": "1.0.0",
    "description": "API documentation for the E-Commerce Platform",
    "contact": {
      "name": "Mohamed Fathy",
      "email": "moham6dfathy@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Category": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "parent": { "type": "string" }
        }
      },
...
    }
  },
  "paths": {
...
    "/categories": {
      "get": {
        "tags": ["Categories"],
        "summary": "Get all categories",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of categories",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Category" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Categories"],
        "summary": "Create a category",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Category" }
            }
          }
        },
        "responses": { "201": { "description": "Category created" } }
      },
      "delete": {
        "tags": ["Categories"],
        "summary": "Delete all categories",
        "security": [{ "bearerAuth": [] }],
        "responses": { "204": { "description": "All categories deleted" } }
      }
    },
...
  }
}